refactor : validator 하위 패키지 enums 추가 (#559)
