+++ src/main/java/com/clubber/ClubberServer/global/validator/enums/EnumValidator.java
@@ -1,5 +1,6 @@
-package com.clubber.ClubberServer.global.validator;
+package com.clubber.ClubberServer.global.validator.enums;
 
+import com.clubber.ClubberServer.global.validator.enums.Enum;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
